Triangle insertion

+++ app/tetwild/Parameters.h
@@ -9,13 +9,13 @@ struct Parameters
     double lr = 5e-2; // target edge length (relative)
     double l = -1.;
     double diag_l = -1.;
-    Vector3f min = Vector3f::Zero();
-    Vector3f max = Vector3f::Ones();
+    Vector3d min = Vector3d::Zero();
+    Vector3d max = Vector3d::Ones();
 
     double splitting_l2 = -1.; // the lower bound length (squared) for edge split
     double collapsing_l2 = std::numeric_limits<double>::max(); // the upper bound length (squared) for edge collapse
 
-    void init(const Vector3f& min_, const Vector3f& max_)
+    void init(const Vector3d& min_, const Vector3d& max_)
     {
         min = min_;
         max = max_;
